+++ src/Connection.php
@@ -7,10 +7,11 @@ namespace Yiisoft\Db\Sqlite;
 use Yiisoft\Db\Driver\Pdo\AbstractPdoConnection;
 use Yiisoft\Db\Driver\Pdo\PdoCommandInterface;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
+use Yiisoft\Db\Schema\Column\ColumnFactoryInterface;
 use Yiisoft\Db\Schema\Quoter;
 use Yiisoft\Db\Schema\QuoterInterface;
 use Yiisoft\Db\Schema\SchemaInterface;
-use Yiisoft\Db\Sqlite\Column\ColumnBuilder;
+use Yiisoft\Db\Sqlite\Column\ColumnFactory;
 use Yiisoft\Db\Transaction\TransactionInterface;
 
 use function str_starts_with;
@@ -71,9 +72,9 @@ final class Connection extends AbstractPdoConnection
         return $this->queryBuilder;
     }
 
-    public function getColumnBuilderClass(): string
+    public function getColumnFactory(): ColumnFactoryInterface
     {
-        return ColumnBuilder::class;
+        return new ColumnFactory();
     }
 
     public function getQuoter(): QuoterInterface