+++ tests/ConnectionTest.php
@@ -14,7 +14,7 @@ use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Profiler\ProfilerInterface;
-use Yiisoft\Db\Sqlite\Column\ColumnBuilder;
+use Yiisoft\Db\Sqlite\Column\ColumnFactory;
 use Yiisoft\Db\Sqlite\Tests\Support\TestTrait;
 use Yiisoft\Db\Tests\Common\CommonConnectionTest;
 use Yiisoft\Db\Transaction\TransactionInterface;
@@ -182,11 +182,11 @@ final class ConnectionTest extends CommonConnectionTest
         $this->assertTrue($thrown, 'An exception should have been thrown by the command.');
     }
 
-    public function testGetColumnBuilderClass(): void
+    public function testGetColumnFactory(): void
     {
         $db = $this->getConnection();
 
-        $this->assertSame(ColumnBuilder::class, $db->getColumnBuilderClass());
+        $this->assertInstanceOf(ColumnFactory::class, $db->getColumnFactory());
     }
 
     private function createProfiler(): ProfilerInterface