+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Enh #315: Raise minimum PHP version to `^8.1` with minor refactoring (@Tigrov)
 - Enh #314: Implement `ColumnFactory` class (@Tigrov)
 - Enh #317: Separate column type constants (@Tigrov)
+- Enh #318: Realize `ColumnBuilder` class (@Tigrov)
 
 ## 1.2.0 March 21, 2024